+++ CHANGELOG.rst
@@ -2,6 +2,12 @@
 Changelog
 =========
 
+0.2.3 (2917-05-02)
+==================
+
+- Add ``full_name`` property to ``Setting`` class.
+- Add ``required`` parameter to ``Setting`` class (default ``False``).
+
 0.2.2 (2017-04-17)
 ==================
 
@@ -10,7 +16,7 @@ Changelog
 0.2.1 (2017-04-17)
 ==================
 
-- Add PositiveInt and PositiveFloat settings.
+- Add ``PositiveInt`` and ``PositiveFloat`` settings.
 - Add support for Django 1.11.
 - Implement basic settings classes.
 
@@ -19,7 +25,7 @@ Changelog
 
 - Implement basic Setting class.
 - Pin dependencies.
-- Change distribution name to app-settings.
+- Change distribution name to ``app-settings``.
 
 0.1.0 (2017-03-23)
 ==================